Replace volume for fcos-1
I've created a new volume for fcos-1 and moved all stuff
over before removing the old volume.

We had a lot of issues with the previous volume, where
we typically see max 10 read IOPS.

Note that the volume has volume type rbd which become the
default when I created it similar as previously. I did try to
create another volume with the "correct" volume type but it had
the same IOPS issues.